skills/boron-nmr-predict/requirements-core.txt
numpy
pandas
scikit-learn
torch==2.3.1
rdkit-pypi
Pillow
matplotlib
huggingface_hub
